+++ tiger-testenv-mgr/src/frontend/src/components/replay/ScenarioReplayer.ts
@@ -5,7 +5,7 @@
 import ScenarioIdentifier from "@/types/testsuite/ScenarioIdentifier";
 
 export function replayScenario(scenarioIdentifier: ScenarioIdentifier) {
-  fetch(process.env.BASE_URL + "replay", {
+  fetch(import.meta.env.BASE_URL + "replay", {
     headers: new Headers({'content-type': 'application/json'}),
     method: "POST",
     body: JSON.stringify(scenarioIdentifier)