+++ ReleaseNotes.md
@@ -15,6 +15,7 @@
 * TGR-1241: Added Case-Insensitive matching in RbelPath-Expressions: `$.body.[~'bar']`
 * TGR-1309: CBOR added to Rbel
 * TGR-1298: Updated frontend libraries.
+* TGR-1310: migrated frontend build tool from vue-cli to vite.
 
 ## Bugfixes